Fix jobs rollback
The jobs table is used for queuing - there should be no active jobs when running this. i.e. the jobs table should be empty.
The most reliable way to rollback is simply to drop and recreate the previous version of the jobs table. 
This avoids issues with not-nullable and tinyInteger in an existing table encountered when just restoring the "reserved" column.