New g2f_r container
Here we split the r and python scrips from the g2f container to two containers, g2f for python scripts, and g2f_r for R scripts. The reason why we did this is that because we needed to update R to 4.2.3 but by doing that we have many conflicts with the obsolete python 2 that must be used in the g2f container.