New JobHelper kill_docker_containers_based_on_jobID
On job timeout or other undefined failure the failed() method is being called.
This method is responsible for the cleanup after these events.
The cleanup consists of killing the remaining containers.
That's why we call the JobHelper::kill_docker_containers_based_on_jobID that takes the jobID as input, searches for containers that have this id in their names and kills them one by one.

To do that we needed to replace all exec functions with laravel's built-in functionality of processes, so that, a job can fail on timeout without the need to wait for all its sub-processes to finished.