new EnsureJobBelongsToLoggedInUser middleware
In order to wrap all the necessary web routes using this middleware we needed to rename all the id variables to jobID, for the middleware to be able to read this var in a universal way and check if the job belongs to the logged in user. Otherwise throw an error.