input.gwas file check is only necessary for snp2gene
Check the existence of that file only when the job type is snp2gene and not when it is geneMap.

GeneMap does not need the input file to run. It uses intermediate files.

input params file is still needed for both types of jobs, that's why we check it at the very beginning.