Move page globals to pageStateComponents
These singelton stores can be imported.
This make the use of page globals explicit
and well defined.
TODO: It would be better to refactor to remove
page globals. There needs to be a clear distinction
between reusable functions that have all state in parameters
and functions that are tied to a particular page and
its global information.